+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+- Added support for icons in breadcrumb items via the `icon` property.
+- Added interactive dropdown for truncated items (when using `maxItems`), allowing users to access hidden links.
+- Added `activeLink` prop to optionally render the current page item as a clickable link.
+
 ### Dependencies
 - Updated the `@todovue/tv-demo` dependency to `^1.4.4`.
 - Updated the `sass` dependency to `^1.97.2`.